// src/types.d.ts
export type StateName = string;
export type EventName = string;
export type EventPayload = any;
export type OnEnter = (meta: StateChangeHookMeta, misc: {handlePrivate: HandlePrivate}) => any;
export type OnExit = (meta: StateChangeHookMeta) => any;
export type StateChangeHookMeta = {
  fromStateName: StateName,
  toStateName: StateName,
  eventPayload: EventPayload,
  changeStatePayload: ChangeStatePayload,
};
export type ChangeStateClosure = (toStateName: StateName, changeStatePayload?: ChangeStatePayload) => void;
export type ChangeStatePayload = any;
export type Handler = (
  changeState: ChangeStateClosure,
  eventPayloadObj: { eventPayload: EventPayload },
  misc: {
    handlePrivate: HandlePrivate
  }
) => any;
export type HandlerObj = {
  fn: Handler;
  isPrivate: boolean;
};
export type HandlePrivate = (eventName: EventName, eventPayload: EventPayload) => Promise<HandleMeta>;
export type HandlerMiddleware = (meta: any) => any;
/** A custom setter to write the state, used if state is stored elsewhere (such as redux store or DB) */
export type WriteActiveStateName = <TStateName extends StateName>(stateName: TStateName) => TStateName;
/** A custom getter to read the state, used if the state is stored elsewhere (such as redux store or DB) */
export type ReadActiveStateName = () => StateName;

export type EventMeta = {
  activeStateName: StateName;
  eventName: EventName;
  eventPayload: EventPayload;
  isPrivate: boolean;
  stateChangeCountSnapshot: number;
};

export type StateChangeMeta = {
  activeStateName: StateName;
  toStateName: StateName;
  fromStateName: StateName;
  nextState: State;
};

export type StateChangeHandler = (meta: HandleMeta) => void;

export type HandleStateChangeResult = {
  fromStateName: StateName;
  toStateName: StateName;
  enter: ReturnType<State['enter']>;
  exit: ReturnType<State['exit']>;
};

export type HandleMeta = {
  hasHandler: boolean;
  beforeHandleResult: ReturnType<HandlerMiddleware>;
  afterHandleResult: ReturnType<HandlerMiddleware>;
  handlerResult: ReturnType<Handler>;
  changeStateResult: HandleStateChangeResult;
  isPrivate: boolean;
};

/**
 * Constructor configuration object for StateMachine class.
 */
export type StateMachineConfig = {
  /** A custom setter to write the state, used if state is stored elsewhere (such as redux store or DB) */
  writeActiveStateName: WriteActiveStateName;
  /** A custom getter to read the state, used if the state is stored elsewhere (such as redux store or DB) */
  readActiveStateName: ReadActiveStateName;
};

/**
 * The namespace for the StateMachine class.
 */
export type StateMachinePrivateNamespace = {
  /** The currently active state name. */
  activeStateName: StateName;
  /** The mapping reference of states. */
  states: {[KStateName in StateName]: State};
  /** The state machine's event emitter. */
  emitter: EventEmitter;
  /** A custom setter to write the state, used if state is stored elsewhere (such as redux store or DB) */
  writeActiveStateName: WriteActiveStateName;
  /** A custom getter to read the state, used if the state is stored elsewhere (such as redux store or DB) */
  readActiveStateName: ReadActiveStateName;
  /** How many times the state has changed. Acts as an identifier for the current state change. */
  stateChangeCount: number;
  /** Backlogged events to dispatch. */
  pendingDispatchEventMeta: EventMeta[];
};